+++ temba/orgs/tests.py
@@ -1,7 +1,6 @@
 import io
 import smtplib
 from datetime import date, datetime, timedelta
-from decimal import Decimal
 from unittest.mock import patch
 from urllib.parse import urlencode
 
@@ -17,11 +16,10 @@ from django.urls import reverse
 from django.utils import timezone
 
 from temba import mailroom
-from temba.airtime.models import AirtimeTransfer
 from temba.api.models import APIToken, Resthook, WebHookEvent
 from temba.archives.models import Archive
 from temba.campaigns.models import Campaign, CampaignEvent, EventFire
-from temba.channels.models import Alert, Channel, ChannelLog, SyncEvent
+from temba.channels.models import Channel, ChannelLog, SyncEvent
 from temba.classifiers.models import Classifier
 from temba.classifiers.types.wit import WitType
 from temba.contacts.models import (
@@ -37,6 +35,7 @@ from temba.flows.models import ExportFlowResultsTask, Flow, FlowLabel, FlowRun, 
 from temba.globals.models import Global
 from temba.locations.models import AdminBoundary
 from temba.msgs.models import ExportMessagesTask, Label, Msg
+from temba.notifications.incidents.builtin import ChannelDisconnectedIncidentType
 from temba.notifications.types.builtin import ExportFinishedNotificationType
 from temba.request_logs.models import HTTPLog
 from temba.schedules.models import Schedule
@@ -757,6 +756,7 @@ class OrgTest(TembaTest):
         assert_org(self.org2, is_suspended=False)
 
         self.assertEqual(1, self.org.incidents.filter(incident_type="org:suspended", ended_on=None).count())
+        self.assertEqual(1, self.admin.notifications.filter(notification_type="incident:started").count())
 
         self.org.suspend()  # noop
 
@@ -1663,22 +1663,6 @@ class OrgTest(TembaTest):
         # and that we have the surveyor role
         self.assertEqual(OrgRole.SURVEYOR, self.org.get_user_role(User.objects.get(username="beastmode@seahawks.com")))
 
-    def test_has_airtime_transfers(self):
-        AirtimeTransfer.objects.filter(org=self.org).delete()
-        self.assertFalse(self.org.has_airtime_transfers())
-        contact = self.create_contact("Bob", phone="+250788123123")
-
-        AirtimeTransfer.objects.create(
-            org=self.org,
-            contact=contact,
-            status=AirtimeTransfer.STATUS_SUCCESS,
-            recipient="+250788123123",
-            desired_amount=Decimal("100"),
-            actual_amount=Decimal("0"),
-        )
-
-        self.assertTrue(self.org.has_airtime_transfers())
-
     def test_prometheus(self):
         # visit as viewer, no prometheus section
         self.login(self.user)
@@ -2170,11 +2154,7 @@ class OrgDeleteTest(TembaTest):
                 [],
             )
         )
-        add(
-            Alert.objects.create(
-                channel=channel2, alert_type=Alert.TYPE_POWER, created_by=self.admin, modified_by=self.admin
-            )
-        )
+        add(ChannelDisconnectedIncidentType.get_or_create(channel2))
         add(ChannelLog.objects.create(channel=channel1, log_type=ChannelLog.LOG_TYPE_MSG_SEND))
         add(
             HTTPLog.objects.create(