+++ temba/utils/email.py
@@ -128,8 +128,12 @@ def send_temba_email(subject, text, html, from_email, recipient_list, connection
             message.send()
         else:
             send_mail(subject, text, from_email, recipient_list, connection=connection)
-    else:
+    else:  # pragma: no cover
         # just print to console if we aren't meant to send emails
-        print("----------- Skipping sending email, SEND_EMAILS to set False -----------")
+        print("------------- Skipping sending email, SEND_EMAILS is False -------------")
+        print(f"To: {', '.join(recipient_list)}")
+        print(f"From: {from_email}")
+        print(f"Subject: {subject}")
+        print()
         print(text)
         print("------------------------------------------------------------------------")